Added portfolio tracking to service from tools.
