--- tools/portfolio/portfolio_test.go
@@ -1,7 +1,0 @@
-package main
-
-import "testing"
-
-func TestMain(t *testing.T) {
-
-}
